% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy-package.R
\docType{package}
\name{fy-package}
\alias{fy}
\alias{fy-package}
\title{Utilities for financial years}
\description{
A financial year starts halfway through a calendar year.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
