% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_fy.R
\name{next_fy}
\alias{next_fy}
\alias{prev_fy}
\title{Next and previous financial years}
\usage{
next_fy(fy, h = 1L)

prev_fy(fy, h = 1L)
}
\arguments{
\item{fy}{A financial year as a character vector.}

\item{h}{An integer, the "horizon" to go forward (for \code{next_fy}) or
backward (for \code{prev_fy}).}
}
\description{
Next and previous financial years
}
